Change binary book to a new format that doesn't require hash code storage with each move.
Simplify book-handling code and remove support for score and
result learning, which modified the binary book file (this was
a complication, lost info when the book was re-built, and poses
problems in a multi-user OS when write access to the book is not
guaranteed). Position learning, aka permanent brain, is still
supported, since this uses an external text file.