Fix bogus PV updates in quiesce()
node->best is not necessarily the most recently analyzed move there, so
it can't be assumed that node+1 contains a continuation from node->best.

Just drop moves from recursive calls to quiesce() rather
than trying to keep track of them. It's the very tip of the search tree
so the moves are probably not good anyway.