There were two classes tracking move history: the Log class and the MoveArray class. Besides handling writes to the log file, the history in the Log class was used for position learning. The MoveArray class was used for repetition detection and for supporting PGN game output.
This commit removes the Log class and consolidates information
necessary for repetition tracking and learning into MoveArray, which
now derives from std::vector. Log output to a file is no longer
supported, but logging is still available through interfaces
supporting Winboard or UCI debug messages.

In addition, this commit enhances move history tracking so that it is
accurate in UCI mode, enabling position learning to be supported under
UCI.